Implementant el sistema de seguritat

+++ src/Controller/SongController.php
@@ -26,6 +26,9 @@ class SongController extends AbstractController
     #[Route('/', name: 'app_song_index', methods: ['GET'])]
     public function index(Request $request, PaginatorInterface $paginator, SongRepository $songRepository): Response
     {
+        $this->denyAccessUnlessGranted('ROLE_ADMIN',
+            null, 'Accés restringit, soles adminis');
+
         // Accedir a tots els àlbums (necessita propietat i constructor per accedir al entityManager)
         $albums = $this->entityManager->getRepository(Album::class)->findAll();
 
@@ -54,6 +57,9 @@ class SongController extends AbstractController
     #[Route('/new', name: 'app_song_new', methods: ['GET', 'POST'])]
     public function new(Request $request, EntityManagerInterface $entityManager): Response
     {
+        $this->denyAccessUnlessGranted('ROLE_ADMIN',
+            null, 'Accés restringit, soles adminis');
+
         $song = new Song();
         $form = $this->createForm(SongType::class, $song);
         $form->handleRequest($request);
@@ -82,6 +88,9 @@ class SongController extends AbstractController
     #[Route('/{id}/edit', name: 'app_song_edit', methods: ['GET', 'POST'])]
     public function edit(Request $request, Song $song, EntityManagerInterface $entityManager): Response
     {
+        $this->denyAccessUnlessGranted('ROLE_ADMIN',
+            null, 'Accés restringit, soles adminis');
+
         $form = $this->createForm(SongType::class, $song);
         $form->handleRequest($request);
 
@@ -100,6 +109,9 @@ class SongController extends AbstractController
     #[Route('/{id}/delete', name: 'app_song_delete', methods: ['POST'])]
     public function delete(Request $request, Song $song, EntityManagerInterface $entityManager): Response
     {
+        $this->denyAccessUnlessGranted('ROLE_ADMIN',
+            null, 'Accés restringit, soles adminis');
+
         if ($this->isCsrfTokenValid('delete'.$song->getId(), $request->getPayload()->get('_token'))) {
             $entityManager->remove($song);
             $entityManager->flush();
